+++ src/action/helper.js
@@ -3,6 +3,8 @@ export const GET_CAROUSEL = "GET_CAROUSEL"
 export const GET_CATEGORY = "GET_CATEGORY"
 
 // USER
+export const VERIFY = "VERIFY"
+export const VERIFY_ERROR = "VERIFY_ERROR"
 export const LOG_IN = "LOG_IN"
 export const LOG_IN_ERROR = "LOG_IN_ERROR"
 export const REGISTER = "REGISTER"
