+++ src/app.jsx
@@ -5,8 +5,8 @@ import Navbar from "./component/navbar"
 
 import Home from "./pages/homepage"
 import Category from "./pages/categoryPage"
-import LoginRegister from "./pages/loginRegister"
 import Register from "./pages/register"
+import Login from "./pages/login"
 import Account from './pages/account'
 import Verification from './pages/verification'
 
@@ -16,8 +16,8 @@ const App = () =>{
             <Navbar/>
             <Route path="/" component={Home} exact/>
             <Route path="/Kategori" component={Category}/>
-            <Route path="/Login-register" component={LoginRegister} />
             <Route path="/Register" component={Register} />
+            <Route path="/Login" component={Login} />
             <Route path="/Akun" component={Account} />
             <Route path="/Verifikasi" component={Verification} />
         </div>
